Fixed math bug in Affine.gen_key()
The code in Hacking Secret Ciphers with Python says to generate the add_key with random.randint(1, len(self.CHARS)), but this is inclusive; so eventually an add_key was generated that was equal to len(self.CHARS), which has the effect of incrementing mult_key by 1. This fucked things. 

Things now unfucked.